Moved XUnitLogger and co to their own project
