+++ src/Common.props
@@ -18,6 +18,7 @@
     <AppMotorCoreVersion>0.14.1</AppMotorCoreVersion>
     <AppMotorCliAppVersion>0.7.1</AppMotorCliAppVersion>
     <AppMotorHttpServerVersion>0.3.1</AppMotorHttpServerVersion>
+    <AppMotorTestCoreVersion>0.1.0</AppMotorTestCoreVersion>
   </PropertyGroup>
 
 </Project>
