+++ tests/AppMotor.HttpServer.Tests/AppMotor.HttpServer.Tests.csproj
@@ -10,6 +10,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\AppMotor.HttpServer\AppMotor.HttpServer.csproj" />
+    <ProjectReference Include="..\..\src\AppMotor.TestCore\AppMotor.TestCore.csproj" />
     <ProjectReference Include="..\AppMotor.Core.Tests\AppMotor.Core.Tests.csproj" />
   </ItemGroup>
